Ceci n'est pas un commit operationnel (ne marche pas). Mais certains bugs ont été corrigés à savoir: - space_mem->insert()->update() : Ne plus set_cow tous les pages (seulement le UTCB) et certaines améliorations sont en cours: - set_cow à certains pages seulement : cela implique passer le nom du Pd en argument à space_mem::update(). Il faut penser à restreindre à un EC précis. Ou plutôt sélectionner les EC qu'on veut durcir - tous les update() sont entrain d'être remplacer par replace_cow() (c'est pas encore sur de marcher, mais c'est nécessaire car update était lourd et impliquait le TYPE_UP/DN/DF qu'on n'a jamais compris) - le Port IO est entrain d'etre géré autrement. Maintenant, le bitmap n'est plus remplacé (par un FULL_1) qu'à son exécution. Quand l'EC n'est pas ordonancé, son bitmap est normal

+++ include/cow.hpp
@@ -19,13 +19,13 @@ private:
     static Spinlock cow_lock;
 
 public:
-    static uint16 frame_index_max;
-    static uint16 elt_index_max;
+    static uint32 frame_index_max;
+    static uint32 elt_index_max;
     static bool max_displayed;
 
     struct cow_frame {
         Paddr phys_addr;
-        uint16 index;
+        uint32 index;
         bool used = false;
     };
     static struct cow_frame cow_frames[NB_COW_FRAME];
@@ -64,7 +64,7 @@ public:
     virtual ~Cow();
 
     static struct block* get_new_block_elt() {
-        for (uint16 i = 0; i < NB_BLOCK_ELT; i++) {
+        for (uint32 i = 0; i < NB_BLOCK_ELT; i++) {
             if (!block_elts[i].used) {
                 block_elts[i].used = true;
                 return &block_elts[i];
@@ -77,8 +77,8 @@ public:
      * @param elt_index
      * @return 
      */
-    static bool get_cow_list_elt(uint16 & elt_index);
-    static bool get_new_cow_frame(uint16 & frame_index);
+    static bool get_cow_list_elt(uint32 & elt_index);
+    static bool get_new_cow_frame(uint32 & frame_index);
     static void free_cow_frame(struct cow_frame * frame_ptr);
 
     static void free_cow_elt(cow_elt * cow);
